requirements.txt
Django==6.0.3
psycopg2-binary==2.9.11
pillow==12.1.1
django-filter==25.2
sqlparse==0.5.5
asgiref==3.11.1
django-cors-headers
django-extensions
gunicorn
whitenoise
dj-database-url
python-dotenv
